refactor(design): simplify meshing facade
